Add rspec-sidekiq gem
