+++ Gemfile
@@ -57,6 +57,7 @@ group :test do
   gem "faker"
   gem "rails-controller-testing"
   gem "rspec-rails"
+  gem 'rspec-sidekiq'
   gem "simplecov"
   gem "timecop"
   gem "webmock"
